Bug#35089304: Incorrect grant is used if account is specified with an IP address
When comparing two grants to sort them the server would short-circuit
the sort code if the two grants were done for a user with an IP address
instead of a host name and would compare them the same if the IP mask
matches.
Fixed by using the mask only if the masks of the two grants are
different and reverting to the sort value (as is done for host names)
when they are the same.

Change-Id: Idc1307dbcc984decb23ba693d1578f7b670e3080