BUG#34777981: FIX SECONDARY_LOAD FLAG AFTER DDL
Problem:
After a DDL is executed on a table loaded to HeatWave, for a short
period of time, the SECONDARY_LOAD metadata flag is found reset on the
data dictionary. If there is a MySQL crash or a failover in MySQL HA,
the table will not get automatically reloaded on HeatWave Recovery.

RCA:
When alter table executes, it copies the properties that aren't modified
by the alter statement from the existing table definition into a new
table definition. And the new table definition is stored for the altered
table. In this process the secondary_load information was never copied.

Earlier, as DDLs were not allowed on HeatWave tables, it never caused an
issue. However, now DDLs are allowed on Heatwave tables. So, right after
the DDL, the table is marked as STALE in HeatWave. And, until HeatWave's
recovery job reloads the table back (and sets the metadata correctly),
the secondary_load metadata information is missing from the table's
definition. In this small window, if any operation is executed that
relies on the metadata, it would not function properly.

Solution:
In this commit, new member secondary_load is introduced to TABLE_SHARE
and HA_CREATE_INFO. TABLE_SHARE::secondary_load information is populated
from the data-dictionary.
While altering the table, TABLE_SHARE::secondary_load is copied to the
HA_CREATE_INFO::secondary_load to store secondary_load value in data-
dictionary properly. This ensures metadata is always kosher even if
the table is momentarily unloaded from HeatWave memory.

Change-Id: I35f8b34835c16dafee2bb10155aac6a6361ede6c