Bug#35180841 Remove requirement for md5_hash() input to be 8-byte aligned
As described in bug report, there seems to be no valid reasons for
requiring the md5_hash input to be 8-byte aligned anymore.

It seems to just add complexity for the caller, and sometimes
an extra memcpy() may even be needed.

Patch replace the Uint64 accesses in md5_hash with memcpy() instead,
which seems to perform identical to the original DWORD assignments.
(And generate ~identical code as well)

md5_hash() signature is also changed to take a char* instead of
an Uint64*. Inlined variants of md5_hash, taking an Uint32* as
argument, is also provided for convenience and avoiding type casts
where md5_hash is normally called with a key stored in an Uint32[]

Usage of md5_hash is adapted accordingly.

Allocation of temporary buffers for md5_hash as Uint64[] has been
changed to Uint32[] as the Uint64 alignment requirements was removed.

Change-Id: I6089a3af1436d1e73620fabf01b32082519d9b38