Bug#35155005 Buffers for holding a key of 'MAX_KEY_SIZE' are allocated too small
NdbIndexScanOperation::getDistKeyFromRange() allocated the tmp[] buffer as a working area for
Ndb::computeHash() to xfrm the key into if need.

The buffer was allocated with the correct size for transforming a key
of NDB_MAX_KEY_SIZE, multiplied with the MAX_XFRM_MULTIPLY.

However, the same buffer was also used in case a key column
needed shrink_varchar()-reformat (and copy) of the key. Thus,
possibly leaving a too small buffer remaining for the hash-xfrm
of the key in ::computeHash()

Patch divide the buffer usage into two seperate buffers:

  - Uint32 xfrmbuf[ MAX_KEY_SIZE_IN_WORDS * MAX_XFRM_MULTIPLY ]
    Used only as argument to ::computeHash()

  - Uint32 shrinkbuf[ MAX_KEY_SIZE_IN_WORDS ]
    Used in case we need to shrink_varchar()-copy the key.

Change-Id: I9589269ce0e2dc084f262817f0708797032a9e84