Bug#34727136 gr_applier_early_failure failed on daily-8.0
Problem
=======================

gr_applier_early_failure and gr_recovery_single_server fail
on regression testing and in ASAN with heap-use-after-free
in cleanup_secure_connections_context().

Analysis and Fix
=======================
cleanup_secure_connections_context() needs to be called by any thread
that directly or indirectly creates connections when using the XCom
communication stack and a specific OpenSSL version. GR ends up creating
connections indirectly when it starts GCS, in the join process. Thus it
is mandatory that it calls cleanup_secure_connections_context().

cleanup_secure_connections_context() exists as part of a Network
Provider interface. Each Network Provider provides an implementation
of that method in order to remove any resources used by OpenSSL.
The problem is that cleanup_secure_connections_context() is called in
Gcs_interface_factory::cleanup_thread_communication_resources(),
which is the very end of GCS execution. Any instance of a Network
Provider, such as Xcom_network_provider, does not exist anymore.

The solution is:
- Create a new method named delayed_cleanup_secure_connections_context()
  in the Network Provider interface, that returns a Callable
  (std::function) with the code that allows the cleanup to happen on
  an arbitrary time. The Callable must only contain static code.
- Whenever we start and stop the Active Provider on the Network Provider
  Manager, we store that Callable for later use.
- Create a new method in the Network Provider Manager that allows
  the stored Callable to be called from any poing in the code.

Using this approach, we guarantee that we always have the cleanup code
available and in a generic way, provided by any implementation of the
Network Provider interface.

Change-Id: Ibbabc0ccb8cb06abe1135dcc30e642faca82185f