Bug#34823462 [InnoDB] Assertion !user_table->drop_aborted (prepare_inplace_alter_table_dict)
If an index creation is interrupted by 'kill query', InnoDB
attempts to drop the index that was being created, unless
another thread references the table, in which case it merely
sets a flag called `drop_aborted`. This flag should signal
the other thread, to drop the specific aborted index. But
when the other reference is held by the stats background thread,
that bg thread might leave the index as is. This can trigger the
assert in the title of the bug at a later point, during another
DDL operation.

An example of the sequence of events when this happens:

connection1
let $ID= `SELECT @id := CONNECTION_ID()`;
CREATE UNIQUE INDEX idx1  ON t1;

connection2
KILL QUERY @id;
-- index creation interrupted during Builder::btree_build()
-- background stats thread calls dict_table_stats_lock(table, RW_X_LATCH) in dict_stats_update_persistent (without this patch)
-- the alter table thread intends to drop the aborted index, but since there is a reference by another thread, it doesn't drop it
-- background stats threads unlocks the table, without dropping the aborted index

The fix is to make to wait for the background stats thread to finish
working, if it is running. Calling dict_stats_wait_bg_to_stop_using_table
makes sure, that the background thread doesn't start working on the same
table without waiting for the data dictionary lock.

From the documentation of dict_stats_wait_bg_to_stop_using_table:
"The background stats thread is guaranteed not to start using the specified
table after this function returns and before the caller unlocks the data"

Change-Id: I691c75b30e6ebcd7705b49ee7eab35e8764e3f20