Bug#35189721 connection-sharing fails after change-user
Connection Sharing requires access to the user's password to
reauthenticate the connection.

When change-user is called by the user, the stored password gets reset
and connection-sharing gets disabled.

If connection-sharing is enabled, the authentication flow should try
to get the users password via caching-sha2-password's
perform-full-auth flow.

Change
------

If connection-sharing is enabled in the config,

- and the server says "switch-to native_password" and the client supports
  "caching-sha2-password", use caching-sha2-password with the client,
  get the password and use native-password with the server.
- and the server says "switch-to caching-sha2", forward it to the client
  and get the password from the client try fast-auth with server.

- added tests which check that connection-sharing is possible after
  change-user.

Change-Id: Iccb55ce1480dafed5784048bfeb9cbf944b0781f