Bug#35120119 Hypergraph: rescan_cost for hash-join is too low
A hash join is an equi-join of two tuple streams: the build stream and
the probe stream. Given that the build stream fits in memory
(determined by the join_buffer_size variable), this proceeds as
follows: A hash table is built from the build stream, by hashing the
tuples on the equi-join values. Then the probe stream is scanned and
each tuple is matched against the hash table.

But if the build stream does not fit in memory, there must be a
preparatory stage where both streams are partitioned into files using
some other hash function. Then each pair of build and probe partitions
are joined using the algorithm above.

In the planning phase, a hash join is represented by an AccessPath
object.  The AccessPath::rescan_cost() function provides an estimate
of the cost of iterating over the result for the second and subsequent
times.

In certain cases, the rescan cost of a hash join can be much lower
than the initial cost (i.e. the cost of the first iteration), since we
can reuse the hash buffer. This is not possible if the build stream
depends on data that may change between iterations, such as values
taken from outside the AccessPath sub-tree of the build stream. We
also cannot reuse the hash table if the input stream is too big to fit
in memory.

The latter criteria means that AccessPath::rescan_cost() needs to take
the probability K of the build stream exceeding join_buffer_size into
account, as this determines whether we will reuse or rebuild the hash
table (provided that the other conditions for reuse are fulfilled.)

The existing algorithm estimates K=0 if the expected size of the build
stream is less than 90% of join_buffer_size, and K=1 otherwise. This
sometimes causes rescan_cost() to give estimates that are far too low,
since it is fairly common for row estimates to be more than 10%
inaccurate, especially if the input stream is (the result of) a
complex expression rather than a simple index or table scan.  Also,
note that if we cannot reuse the hash buffer, the rescan cost also
includes the cost of regenerating the build stream, which, depending
on the complexity of the input stream expression, can have any cost.

This commit fixes that problem by changing the way we estimate K into:

K = min(1, (build_input_size_estimate/join_buffer_size)^2)

That way, there is a non-zero probability of exceeding the buffer size
even when the estimated size of the build stream is much smaller. Then
we calculate:

rescan_cost = (1 - K) * cost_when_reusing_hash_table +
                        K * cost_when_rebuilding_hash_table

In addition, this commit changes the cost calculation for nested loop
joins.  As it was, the cost of the first scan of the inner stream was
also calculated using rescan_cost. Now we use 'cost' instead.

Change-Id: I36ae7be2f9959a96e4e1742c99a22bcebf912350