Bug#35181008 Determine STACK_DIRECTION at runtime rather than configure time
Remove the cmake check, and the config variable in (the generated)
my_config.h. Do the stack direction check at startup of the server
instead.

The main motivation for this is that clang15 with AddressSanitizer
turns on detect_stack_use_after_return=true by default.  So for ASAN
builds, stack direction is no longer a compile-time property.

Change-Id: I7ffb8c02d78affe3c2decbb3a012d3419b8ed143