Bug#35120119 Hypergraph: rescan_cost for hash-join is too low
Post-push fix: An MTR result file had to be updated. Two hash-join
AccesPath objects are now rated as different by
CompareAccessPaths(). Prior to fixing this bug, they would be rated as
equal. But the fix means that since one of the paths is estimated to
use a bit less hash buffer space, it has a slightly lower chance of
exceeding join_buffer_size.  Therefore, it also has a slightly lower
rescan_cost, and thus a better overall score.

Change-Id: I8a488956414ed09ee40e206e70959b5b29940ade