Bug#35173907: Assertion `CompatibleTypesForIndexLookup(eq_item, field, right)' failed
The hypergraph optimizer assumes that simple equalities constructed
from a multiple equality always compare items whose types are
compatible for use in index lookups, and has an assertion to verify
the assumption.

For comparisons between TIME columns and DATE/DATETIME/TIMESTAMP
constants, however, a multiple equality is created even though this
combination of types is explicitly blocked for index lookups in
comparable_in_index(). An assertion is therefore raised by the
hypergraph optimizer if an equality predicate compares an indexed TIME
column to a temporal constant which has a date component.

There is not much point in creating a multiple equality for such
equalities mixing different temporal types, so the bug is fixed by not
creating a multiple equality for this combination of types.

Change-Id: Ibd3935f73c351a551831f73cf92245c69930d116