Bug#35154637 NDB DATA NODE FAILURE ndbrequire(blockToMain(rg.m_block) != V_QUERY)
This is a regression in 8.0.32 due to the fix of Bug#34776970 Assertion
failure in SimulatedBlock::assembleFragments().

A check was added to ensure that no fragmented signals are sent to
V_QUERY blocks.

Check did not take into account that if receiver node was not a data
node the block number have a different structure.

And in that case if the wrongly calculated main block happen to have
same number as V_QUERY that has nothing to do with a V_QUERY block.

For the rare cases there a API node or management server node is the
receiver and happen to have such a block number an extra check is added
that fail only if receiver node is a data node and the calculated main
block number is V_QUERY.

Change-Id: I24e88db3e014d9a0b11045f56dd44425e8949076