WL#14808: Row-ID ordered scans in the hypergraph optimizer[1/9 - ROR_intersect_plan]
Preparatory re-factoring

Structure ROR_INTERSECT_INFO is replaced with ROR_intersect_plan. Along
with what ROR_INTERSECT_INFO had, it also has the scans that are part of the
plan. This simplifies code in get_best_ror_intersect().
Since all the ror scans are part of a Mem_root_array now, the existing
function print_ror_scans_arr() could not be used to print the trace.
So introduced a new one. print_ror_scans_arr() will be removed in a later
patch.

Change-Id: Id75824c9491d0eced67743bb0439ba2ec76221a5