WL#14808: Row-ID ordered scans in hypergraph optimizer [5/9, intersect_order]
Preparatory refactoring

find_intersect_order uses std::stable_sort now to sort the indexes. It does
exactly what was previously done i.e pick the index that covers most fields
first. If 2 indexes cover the same number of fields, then selectivity is used.
The next best index would be the index which covers the most fields which are
not already covered previously.
The only test result change is because both the indexes cover the same fields
and have the same selectivity. So optimizer can pick any one to intersect.

Change-Id: I835f1a7db4b1ec80a04cd2b009b08ddfddb98932