WL#15257 Enhanced performance for set operations
Implements hashing using Robin Hood hash (same hashing method as for
hashed JOINs) to de-duplicate set operands of INTERSECT and EXCEPT. The
space used for this is controlled by

        setop_hash_buffer_size (default 256 * 1024 bytes).

which is the same default as for join_buffer_size.  If this space is
exceeded, we spill to disk, cf. the algorithm described in the WL,
using "chunk" files, similar to what is employed for hashed JOIN
overflow. Essentially, this is a divide and conquer approach in which
we spread the rows according to another hash (tertiary hash function)
of the row partitioning the rows, allowing us to process with the
in-memory hash table one partition (chunk pair) at a time.

But even this can fail if the table statistics are not up to date or
row sizes are skewed (secondary overflow).  In this case, we abort the
hashing approach and fall back on using the result set's materialized
temporary table's non-unique index on the primary hash of the row for
de-duplication of the rows.

Asymptotically with large table sizes and (too) small sizes of
setop_hash_buffer_size we need several *sets* of chunk files because we
don't want to have more than a certain number of temporary files open at
a time (i.e. 128). This will eventually be slower than using the key of
the materialized table (InnoDB). The code does *not* detect this and
fall back on using the key of the materialized table when this would be
faster.  To get optimal performance, it behoves the user to ascertain
that the buffer is large enough for the volume of (distinct) rows of in
the left operand. The work log contains measurements that give an
indication of the performance.

Testing the operation of spill to disk is straight forward, just have
enough rows in the left operand, detected by seeing temporary files
being used by a query, cf. using

        SHOW STATUS LIKE '%tmp%';

before and after, or inspecting the optimizer trace for evidence
("spill to disk initiated").  Testing secondary overflow is possible
by error injection at those locations where space overflow may happen,
cf. the debug-only variable

        debug_setop_secondary_overflow_at

If this variable is set, it will inject a secondary overflow event
during hashing. Three integers must be given to indicate where to inject
the overflow:

  a) set index, cf. explanation in comments for class SpillState
  b) chunk index
  c) row number

syntax: <set-idx:integer 0-based> <chunk-idx:integer 0-based> <row_no:integer 1-based>
Example:

      SET SESSION debug_setop_secondary_overflow_at = '1 5 7';

If the numbers given are outside range on the high side, they will
never trigger any secondary spill. Secondary overflow can also be seen
in the optimizer trace ("spill handling overflow, reverting to
index").

By default, INTERSECT and EXCEPT use hashing with this patch. This may
be disabled in favour of the old approach (i.e. tmp table key) for
testing purposed (e.g. to compare results), via the optimizer switch

        hash_set_operations {true|false}

We also added optimizer trace so the user can detect spill actions, cf.
tests for this in query_expression.test.

Change-Id: I364b29f515cacd2c8f70f73d34b2fd1e0b8de190