Bug#35195459: main.ps and main.sp fail under ASan with hypergraph
The hypergraph optimizer could fail with heap-use-after-free when
planning a query containing a derived table with a UNION, if the query
was part of a prepared statement or a stored procedure.

The problem was that Query_expression::m_root_access_path was not
reset between executions of prepared statements or stored procedures,
so the hypergraph optimizer's call to
tl->derived_query_expression()->root_access_path() could in some cases
return a pointer to an AccessPath allocated and freed in the previous
execution.

Fixed by resetting the root access path of the Query_expression
between each execution.

Change-Id: I964740f0cc391abfd41dd027b7da22a11bd90da9