WL#14808: Row-ID ordered scans in the hypergraph optimizer [7/9, ROR-Union]
Add support for ROR-Union scans in hypergraph optimizer.

An ROR-Union plan is proposed when optimizer detects that all range
scans that are part of an OR condition (SEL_IMERGE), are rowid ordered
scans. A regular index merge (sort and de-duplicate) scan is not proposed
in this case since it would always be cheaper without the sort. It would
have been okay to propose both variants. However the current (existing
old optimizer's) cost model makes ROR-Union costlier than a regular
index merge scans. Since hypergraph optimizer is using the same cost
model for now, an ROR-Union is never picked. So hypergraph optimizer
imitates what is being done in the old optimizer which is to propose
only ROR-Union scan if it is possible to do so.
Handling of applied predicates and subsumed predicates for the scan
is similar to a regular index merge scan. Same with proposing plans
with ordering.

Change-Id: I797e798deae4e9bca1d3dc6bd3931d0c070f9dca