WL#14808: Row-ID ordered scans in the hypergraph optimizer [8/9 - ROR-Intersect]
An ROR-Intersect plan is proposed when there are atleast two range scans
that are rowid ordered. Similar to old optimizer, it first orders the
range scans based on the number of fields that are covered by an index.
This way when a plan is proposed, the best indexes are always looked at
first. Old optimizer could propose only one plan. However hypergraph
optimizer can propose more. This would mean, it could choose any
combination of indexes to do an ROR-Intersect and propose it. But it is
not needed to do so. If adding a rowid ordered scan to an existing
combination of scans does not improve selectivity, Or if fields
covered by a scan are already covered, it does not have to be
proposed. So, what it does is, for each combination of the
available ror-scans, it proposes the plan only if all
the scans used in the combination could be added to a plan. This
way, we do not propose plans that anyways would be rejected later.
For handling the subsumed and applied predicates, we use the same
information that is collected when proposing all possible range scans
earlier. Ordering is similar to that of other index merge scans
(only primary key ordering).

Change-Id: Ib85799884a2b9d7b033ed7353826c5124418fe64