WL#14808: Row-ID ordered scans in the hypergraph optimizer [9/9, ROR-Union with ROR-Intersect]
Propose ROR-Union scans with ROR-Intersect scans. When proposing ROR-Union
plan, it checks if an ROR intersect plan could be proposed instead of one
single range scan to handle one part of an OR condition. While doing this,
it only picks the best ROR-Intersect plan instead of proposing all possible
combinations of ROR-Intersect scans. We use get_best_ror_intersect()
currently used by old optimizer to pick the best ROR-intersect plan.
Subsumed predicates are not handled for these plans as a range tree is
marked inexact if it has more than one predicate in one part of an OR
condition. This could be handled. But it is left as is for now.

Change-Id: Ic70fb0d6b714cede619a4dd4501f61733ba65bc9