Bug#35181739 Valgrind: Conditional jump or move depends on uninitialised value(s)
The srv_sys_space.open_or_create is not initializing the flushed_lsn
argument in case of I/O error - which is to be expected, as in general
it should read this value from system tablespace file.

However, the caller was not checking for error before accessing
flushed_lsn and comparing it LOG_START_LSN. This was wrong.

Also, the caller forgot to call srv_init_abort in case of error.

This could be demonstrated on Linux using very simple scenario, in which
we try to start another instance of mysqld from the same data-dir, which
fails when (on Linux) we try to lock the ibdata1 file.

--let $SEARCH_FILE=$MYSQL_TMP_DIR/another_instance.err
--let $MYSQLD_ARGS=--datadir=$MYSQLD_DATADIR --log-error=$SEARCH_FILE

--echo $MYSQLD_CMD $MYSQLD_ARGS
--exec valgrind --tool=memcheck $MYSQLD_CMD $MYSQLD_ARGS

As this test requires: Linux, manual usage of valgrind, messing with
original data-dir, and would still complain about leaks during abort,
and most importantly - doesn't really do anything meaningful after the
bug is fixed - it makes no sense to add it to the suite.

Change-Id: I24ab568cf81c0540a3f503b2e0dc3563117a06c1