Bug#35206138: Assigning json_value output to user defined variable will return NULL or empty string if input is empty string
When setting a user-defined variable to the result of a call to
JSON_VALUE, and JSON_VALUE returns an empty string, the value of the
user-defined variable may incorrectly become NULL.

The reason is that JSON_VALUE returns a String whose ptr() returns
nullptr. While there's not really anything wrong with that, as long as
String::length() returns zero, there are some code paths that expect
String::ptr() to return non-nullptr. And Item_func_set_user_var is
among those, and conflates the empty string and NULL if ptr() returns
nullptr.

This patch fixes the bug at both ends. JSON_VALUE is changed so that
it returns Item::make_empty_result() when the result is an empty
string. That helper function makes sure String("", 0) is returned
instead of String(nullptr, 0), and has been used for similar bugs in
the past. Additionally, Item_func_set_user_var is changed to trust
Item::null_value and not assume the value is NULL just because
String::ptr() returns nullptr, in case there are other functions that
behave in a similar way.

Change-Id: Idec4cb6e8213b7c92ac0c3d1ee2d1af46d125213