Bug#34976442 mysqld crashed
Problem
====================

On an MySQL Innodb cluster deployment which has:
- A virtualization platform of Docker/Kuberenetes;
- 3 nodes working and running

During regular operation, one of the nodes crashes and it is was unable
to start and join the group again. In order to restore the system,
the whole group was restarted and then all members crashed in
get_ipv4_local_addresses within GCS/XCom.

Analysis and Fix
====================

Docker/Kuberenetes create a lot of network interfaces within a pod,
for a virtual network. In this specific case, more than 64 interfaces.

What happened was that XCom allocates an array of 64 interfaces which
is reallocated each time we overflow it. Specifically,
`ifrp` is overflowed when network interface is more than
IFRP_INIT_ARR_SIZE(64). But it is not reallocated because wrong
condition `i==ifrpsize`.

The solution is to fix the wrong reallocate condition.

Change-Id: Ibbf150e4f6a7fc84db8d44193c3f729aba7d7089