Bug#35185670 mysqld crashes on expensive query
Preparation patch in order to allow a unit test of the
FragmentedSectionIterator to later be added. - This patch
only move different SectionIterators from TransporterFacade.*
into the new SectionIterator.hpp/ *.cpp files.

This also allows us to hide the implementation of the methods
in these classes into SectionIterator.cpp. It was previously
exposed in order for the existing unit test to find the implementation
without pulling in lots from libraries etc.

The unit tests are also moved into SectionIterator.cpp, and
the built of it added to thge makefiles such that it now becomes as
standard part of the unit test family tested by Pb2.

Change-Id: I93c0836856757acf5f87c204b573cf76daecf2f3