BUG#35217728 Speedup mtr.pl by clean vardir
When using mtr.pl for development (running individual
tests or parts of suites) the time spent setting up the
environment far exceeds the time to run the actual test.

Most timeconsuming are the steps to create the MySQL
system database (i.e running mysqld --initialize)
and copying the std_data directory into the var
directory. Since the size of std_data grows over
time as more and more stuff are added there, this
step takes increasingly longer.

Fix by improving the --fast mode to clean the var
directory for reuse rather than removing it entirely
and creating it again. This allows the data/ and
std_data/ directories to be kept.

Also make it possible to pass --no-fast to negate any previous --fast
arguments one the command line.

With this patch, the time to run a single simple test is decreased with
13 seconds on standard development machine.

  $ mtr alias
  <snip>
  Checking leftover processes
  Removing old var directory
  Creating var directory '/home/mblaudd/mysql/trunk/bld/mysql-test/var'
  - symlinking 'var' to '/dev/shm/var_auto_iXMr'
  Installing system database
  Using parallel: 1

  ==============================================================================
         TEST NAME                         RESULT  TIME (ms)
  ------------------------------------------------------------------------------
  [ 50%] main.alias                                [ pass ]    758
  [100%] shutdown_report                           [ pass ]
  ------------------------------------------------------------------------------
  The servers were restarted 0 times
  The servers were reinitialized 0 times
  Spent 0.758 of 17 seconds executing testcases

  $ mtr alias --fast
  <snip>
  Checking leftover processes
  Cleaning var directory to save time
  Creating var directory '/home/mblaudd/mysql/trunk/bld/mysql-test/var'
  Reusing system database
  Using parallel: 1

  ==============================================================================
       TEST NAME                       RESULT  TIME (ms) COMMENT
  ------------------------------------------------------------------------------
  [ 50%] main.alias                                [ pass ]    696
  [100%] shutdown_report                           [ pass ]
  ------------------------------------------------------------------------------
  The servers were restarted 0 times
  The servers were reinitialized 0 times
  Spent 0.696 of 4 seconds executing testcases

Change-Id: Ibcf82874f61bbf1e1c8746da028045ccbc0a8a50