Bug#33517923 Remove circular include dependencies [noclose]
There are several circular include dependecies in Ndb code. This
typically comes with special requirements on which order to include
diferent files or other special constructs to only include part of a
file.

Most of these circular include dependencies are removed in this patch
by:

- adding forward declaration of classes
- moving include statements from hpp-files to cpp-files
- removing redundant include
- adding includes to remove dependencies on indirect includes
- moving code from always included cpp-file into hpp-file.
- moving inline functions from hpp-file to cpp-file.

Circular dependencies removed are between:

- API.hpp and NdbImpl.hpp
- Backup.hpp and Dblqh.hpp
- Dblqh.hpp and Dbtux.hpp
- NDBT.hpp and NDBT_Tables.hpp
- RWPool.cpp and RWPool.hpp
- TransporterCallback.hpp and TransporterRegistry.hpp
- WOPool.cpp and WOPool.hpp

There are still many indirect include dependencies breaking the
include-what-you-use (IWYU) principle, in this patch only the few needed
to remove circularity was removed.

One instance of circular include dependencies needs more work and is
left as a separate task, this is the circular dependencies between:

- trp_client.hpp, TransporterFacade.hpp, and, ClusterMgr.hpp

Change-Id: I91f6b7f3a2a459ac2e35c1eed76f8a915d4abf63