Bug#35057632: Mysqld server abort from Item_func::eq(Item const*, bool)
This is a regression from Bug#33830659. As part of this
bugfix, when finding rollup item in group by, we started
to use the "rollup_item" set as part of the structure used
for GROUP BY for getting the correct field to handle
temp table fields.
A "rollup_item" for GROUP BY is set when any of the fields
in the query match the grouping expression. However, when
there are duplicate expressions in GROUP BY, the first one
is always used. As a result "rollup_item" for the duplicates
is never set.
For the failing case, when searching for the field to
replace "f2", it goes through the group list and since
"rollup_item" is not set for the second "f1", it fails.
We now check for "rollup_item" before the comparison.

Change-Id: I5a9f297650220a6a93faf31713d6b976bb1eaaeb