Bug#34924127 ddl on fulltext tables created in <= 5.6.19 crashes in dd_drop_fts_table
As a result of bug#18635485 tables for FTS indices created in MySQL
versions 5.6.5-5.6.19 were in system tablespace regardless of
the tablespace of indexed table. After upgrading database to
newer versions of MySQL, when parent table used file-per-table
configuration, FTS tables were also assumed to be also
file-per-table. This caused the server, when dropping the FTS
tables, to also attempt to drop the corresponding tablespace,
which resulted in an assertion. This made it impossible to optimize
or alter the table or to drop the index. The fix checks if the FTS
table is specifically in the system tablespace when it is removed,
and forgoes dropping the tablespace in this case.

Change-Id: If6301047cc2d30d5e58a45816a3e9937a24a49ab