Bug#35185670 mysqld crashes on expensive query
The patch:

A large query, where the 'compiled' query request information exceeded 32K,
need to be sent to the data nodes as a 'fragmented signal'.

When collecting the different parts going into a query request,
we use iterators over the data block to be collected, and copies the
different data pieces into the signal.

For a fragmented signal (only), we use the specific 'FragmentedSectionIterator'
to collect the different pieces going into the query request. That iterator
offers the added functionality of setting up ranges of data to be retrieved,
and reposition to a specific position within the data, such that it can be
collected in multiple 'fragments'

It turned out that the implementation of several methods in the
FragmentedSectionIterator has several flaws:

1) When moving to a position within the data, we also read block of
data from an underlying iterator, possibly reading beyond the 'pos'
to move to. When later fetching the data with getNextWords(), we
attempted to reuse the data already read in advance when positioning.
This logic only worked as intended when all 'remaining' data was
returned in the first (and last) read chunk. Else:
  - If there were more data available from inderlying iterator, the
    already read block was discarded and the next block of data read
    and returned.
  - When we finaly reached the last block in the 'range'
    (LastReadLen == remaining), the last read and returned block
    was returned again for the second time.

Root cause seems to be that that the FragmentedSectionIterator::getNextWords()
is attempting to use the data pointer + len from one call to
getNextWords() on the underlying iterator, *after* it has called
getNextWords() again.

That assumes that the underlying iterator is presenting a direct view on
some kind of stable structure (e.g. Linear or Segmented Section Pointers
etc) rather than an interface to a 'generator' type function which is
e.g. reusing the same buffer memory to provide content.
(Like ReceiverIdSectionIterator).

Note that if we reposition, or set up a range, within a data block just
returned from getNextWord(), the still available, but previously consumed,
data should become available again. This is a common use case when collecting
of fragmented signal slice up the data to be collected.

Patch divide the functionality of the previous member variable 'lastReadPtrLen'
into the new member variables:

 - 'lastReadTotal': Is the size of the data block last read and refererd by
                    'lastReadPtr'
 - 'lastReadAvail': Is the unconsumed size of 'lastReadPtr'

moveToPos() will usually set 'lastReadAvail = lastReadTotal' when it need
to read a new 'fresh' data block. If it reposition within an existing data
block it 'unconsumes' data, thus set 'lastReadAvail <= lastReadTotal'.

Change-Id: I6897ef1974229af290e27891903a3897f4a2a595